Use logger in get_static_renderers for logging
